+++ src/inventory/mod.rs
@@ -1,5 +1,7 @@
 //! Inventory list files and their entries
+mod fields;
 mod item;
 mod list;
+pub(crate) use self::fields::*;
 pub(crate) use self::item::*;
 pub(crate) use self::list::*;